backend/django_server/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.0
djangorestframework-simplejwt>=5.3
python-dotenv>=1.0
psycopg2-binary>=2.9
gunicorn>=21.2
whitenoise>=6.0
pathlib
streamlit
